Extract SubmitButtonComponent
