+++ web/app/components/submit-button.hbs
@@ -1,0 +1,15 @@
+{{#if (or (eq @validation.progress 'waiting') (eq @validation.progress 'running'))}}
+  <button type="button" class="btn btn-danger" disabled={{this.cancel.isRunning}} {{on 'click' (perform this.cancel)}}>
+    Cancel
+  </button>
+{{else}}
+  <div class="d-flex align-items-center gap-3" {{this.tickCurrentTime}}>
+    <button type="button" class="btn btn-primary" disabled={{or (not this.canSubmit) this.submit.isRunning}} {{on 'click' (perform this.submit)}}>
+      Submit
+    </button>
+
+    {{#if this.cannotSubmitReason}}
+      <div class="text-danger">{{this.cannotSubmitReason}}</div>
+    {{/if}}
+  </div>
+{{/if}}
